requirements.txt
appnope==0.1.4
asttokens==3.0.1
blinker==1.9.0
click==8.3.1
comm==0.2.3
cryptography==44.0.1
debugpy==1.8.20
decorator==5.2.1
dnspython==2.8.0
executing==2.2.1
Flask==3.1.3
Flask-SQLAlchemy==3.1.1
iniconfig==2.3.0
ipykernel==7.2.0
ipython==9.10.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.8.0
jupyter_core==5.9.1
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
nest-asyncio==1.6.0
numpy==2.4.2
packaging==26.0
pandas==3.0.1
parso==0.8.6
pexpect==4.9.0
platformdirs==4.9.2
pluggy==1.6.0
prompt_toolkit==3.0.52
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
pymongo==4.16.0
PyMySQL==1.1.2
pytest==9.0.2
Flask-Migrate==4.0.4
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
Flask-JWT-Extended==4.5.3
python-dotenv==1.2.1
pyzmq==27.1.0
six==1.17.0
SQLAlchemy==2.0.47
stack-data==0.6.3
tornado==6.5.4
traitlets==5.14.3
typing_extensions==4.15.0
wcwidth==0.6.0
Werkzeug==3.1.6
